refactor(cli): extract openshell runtime helpers
